examples/python/rtc/pyproject.toml
[project]
name = "livekit-portal-rtc-example"
version = "0.1.0"
description = "Request-driven RTC (real-time chunking) flow on top of Portal."
requires-python = ">=3.12"
dependencies = [
    "livekit-portal",
    "livekit-api>=0.7",
    "python-dotenv>=1",
    "numpy>=1.24",
]

[project.optional-dependencies]
modal = ["modal>=0.73"]

[tool.uv.sources]
livekit-portal = { path = "../../../python/packages/livekit-portal", editable = true }
